==> 1_1_depart_data_by_id.ipynb <==
import pandas as pd

raw_netflow = pd.read_csv("netflow/200702111400.csv")

import os


os.makedirs("netflow/200702111400/", exist_ok=True)
for key, value in raw_netflow.groupby('sa'):
    basename = key.replace(".", "_").replace(":", "_")
    length = value.shape[0]
    os.makedirs("netflow/200702111400/" + str(length // 100 * 100), exist_ok=True)
    value.to_csv("netflow/200702111400/" + str(length // 100 * 100) + "/" + str(length) + "-" + str(basename) + ".csv", index=False)
