==> 1_7_clustering_tskmeans.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import numpy as np

from tslearn.utils import from_pyts_dataset
from tslearn.clustering import TimeSeriesKMeans

INTERVAL = 30

pyts_timeseries = np.load(f"interval{INTERVAL}_reconstructed_STL_trend_ts/200702111400/pyts_dataset.npy")
pyts_timeseries.shape

tslearn_timeseries = from_pyts_dataset(pyts_timeseries)
tslearn_timeseries.shape

model = TimeSeriesKMeans(n_clusters=10, metric="softdtw", max_iter=10, metric_params={"gamma": .01})
model.fit(tslearn_timeseries)
labels = model.labels_

np.asarray(np.unique(labels, return_counts=True)).T

np.save(f"interval{INTERVAL}_reconstructed_STL_trend_ts/200702111400/labels.npy", labels)

# Save model
model.to_pickle(f"interval{INTERVAL}_reconstructed_STL_trend_ts/200702111400/model.pkl")
