==> 2_5_clustering_tskmeans.ipynb <==
import os
import warnings
warnings.filterwarnings('ignore')

import numpy as np

from glob import glob
from pathlib import Path
from tslearn.utils import from_pyts_dataset
from tslearn.clustering import TimeSeriesKMeans

INTERVAL = 5

def dropna(nparray):
    if isinstance(nparray[0], np.ndarray):
        return np.array([dropna(x) for x in nparray])
    else:
        return nparray[~np.isnan(nparray)]

for direction in ["src"]:
    for dirname in glob(Path(f'timeseries/interval_{INTERVAL}_{direction}*').__str__()):
        print("Dir: ", dirname)
        if os.path.exists(os.path.join(dirname, "tslearn_kmeans.npy")):
            continue
        pyts_dataset = np.load(os.path.join(dirname, "pyts_dataset.npy"))
        pyts_dataset = dropna(pyts_dataset)
        print("\tPyts dataset shape: ", pyts_dataset.shape)
        X = from_pyts_dataset(pyts_dataset)
        print("\tTslearn dataset shape: ", X.shape)
        model = TimeSeriesKMeans(n_clusters=10, metric="softdtw", verbose=True, random_state=10, n_jobs=4, max_iter=10, tol=1e-3)
        y_pred = model.fit_predict(X)
        np.save(os.path.join(dirname, "tslearn_tskmeans.npy"), y_pred)
        print("\tTslearn kmeans shape: ", y_pred.shape)
        del pyts_dataset, model, X, y_pred

# pyts_timeseries = np.load(f"interval{INTERVAL}_reconstructed_STL_trend_ts/200702111400/pyts_dataset.npy")
# pyts_timeseries.shape

# tslearn_timeseries = from_pyts_dataset(pyts_timeseries)
# tslearn_timeseries.shape

# model = TimeSeriesKMeans(n_clusters=10, metric="softdtw", max_iter=10, metric_params={"gamma": .01})
# model.fit(tslearn_timeseries)
# labels = model.labels_

# np.asarray(np.unique(labels, return_counts=True)).T

# np.save(f"interval{INTERVAL}_reconstructed_STL_trend_ts/200702111400/labels.npy", labels)

# Save model
# model.to_pickle(f"interval{INTERVAL}_reconstructed_STL_trend_ts/200702111400/model.pkl")
