==> 2_8_similarity.ipynb <==
import math
import json
import os
import warnings
warnings.filterwarnings('ignore')

import numpy as np
import pandas as pd

from glob import glob
from pathlib import Path

INTERVAL = 30
TARGETFILENAME = 'sorted_rfcm_results_similarity.csv'
SOURCEFILENAME = 'sorted_rfcm_results.csv'

for direction in ['src']:
    for targetDir in glob(Path(f"timeseries_feature/interval_{INTERVAL}_{direction}*/").__str__()):
        print("Dir: ", targetDir, " " * 50)
        # if os.path.exists(f"{targetDir}/{TARGETFILENAME}"):
        #     print(f"{TARGETFILENAME} already exists")
        #     continue
        
        with open(os.path.join(targetDir, 'types.json')) as f:
            types = json.load(f)
        df = pd.read_csv(os.path.join(targetDir, SOURCEFILENAME), index_col=0, dtype=types)
        
        df["avg"] = df.mean(axis=1)
        df.sort_values(by="avg", ascending=False, inplace=True)
        
        first_column = df.pop('avg')
        df.insert(0, 'avg', first_column)
        
        df.to_csv(os.path.join(targetDir, TARGETFILENAME))

# with open('timeseries_feature/interval_30_src_feature/types.json') as f:
#     types = json.load(f)
# df = pd.read_csv('timeseries_feature/interval_30_src_feature/sorted_rfcm_results.csv', index_col=0, dtype=types)
# df

# df["avg"] = df.mean(axis=1)
# df

# df.sort_values(by=['avg'])

# number_entries = len(df)
# number_features = len(df.columns)
# for col in df.columns:
#     cluster_size = df[col].value_counts()
#     print(cluster_size)
#     print(math.e ** -(5 * cluster_size / number_entries))
#     break
