==> 3_0_remove_summary.ipynb <==
import pandas as pd

from glob import glob
from pathlib import Path

for file in glob('pch/netflow/*.csv'):
    netflow = pd.read_csv(file)
    netflow.drop(netflow.tail(len(netflow) - netflow[netflow["ts"] == "Summary"].index.values[0]).index, inplace = True)
    netflow.to_csv(file, index=False)
