==> 3_6_merge_rfcm_results.ipynb <==
import json
import os

import numpy as np
import pandas as pd

from glob import glob
from pathlib import Path

INTERVAL = 60

for direction in ['src']:
    for targetDir in glob(Path(f"timeseries_feature/pch/20230329/interval_{INTERVAL}_{direction}*/").__str__()):
        print("Dir: ", targetDir, " " * 50)
        # if os.path.exists(f"{targetDir}/rfcm_results.csv"):
        #     # os.remove(f"{targetDir}/rfcm_results.csv")
        #     print("rfcm_results.csv already exists")
        #     continue
            
        with open(os.path.join(targetDir.replace("timeseries_feature", "timeseries"), "pyts_dataset_sample.txt")) as f:
            ip = [line.strip() for line in f.readlines()]
        
        results = []
        for dirname in glob(Path(f"{targetDir}/*/").__str__()):
            try:
                print("\t", dirname, " " * 20, end="\r")
                label = np.load(os.path.join(dirname, "rfcm_label.npy"))
                df = pd.DataFrame({"ip": ip, os.path.split(dirname)[1]: label})
                df.set_index("ip", inplace=True)
                results.append(df)
            except FileNotFoundError as e:
                continue
            
        result = pd.concat(results, axis=1)
        result.to_csv(os.path.join(targetDir, 'rfcm_results.csv'))
        
        res = result.dtypes.to_frame('dtypes').reset_index()
        d = res.set_index('index')['dtypes'].astype(str).to_dict()
        with open(os.path.join(targetDir, 'types.json'), 'w') as f:
            json.dump(d, f)
