==> 3_7_sort_labels.ipynb <==
import json
import os

import numpy as np
import pandas as p

from glob import glob
from pathlib import Path

INTERVAL = 15
TARGETFILENAME = 'sorted_rfcm_results.csv'

for direction in ['src']:
    for targetDir in glob(Path(f"timeseries_feature/pch/20230329/interval_{INTERVAL}_{direction}*/").__str__()):
        print("Dir: ", targetDir, " " * 50)
        if os.path.exists(f"{targetDir}/{TARGETFILENAME}"):
            print(f"{TARGETFILENAME} already exists")
            continue
        
        with open(os.path.join(targetDir, "types.json")) as f:
            types = json.load(f)
        df = pd.read_csv(os.path.join(targetDir, 'rfcm_results.csv'), index_col=0, dtype=types)
        
        for col in df.columns:
            df[col].replace({key: value for value, key in enumerate(df[col].value_counts().index)}, inplace=True)
        
        df.to_csv(os.path.join(targetDir, TARGETFILENAME))

# with open('timeseries_feature/interval_30_src_feature/types.json') as f:
#     types = json.load(f)
# df = pd.read_csv('timeseries_feature/interval_5_src_feature/rfcm_results.csv', index_col=0, dtype=types)
# df

# for col in df.columns:
#     print(df[col].replace({key: value for value, key in enumerate(df[col].value_counts().index)}))
#     break
